Move all components into "components" folder
