+++ components/Link.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react'
 import NextLink, { LinkProps } from 'next/link'
-import { createLinkStyles } from '../../styles/mixins'
+import { createLinkStyles } from '../styles/mixins'
 import { css, jsx, InterpolationWithTheme } from '@emotion/core'
 /** @jsx jsx */ jsx
