+++ components/PostListItem.tsx
@@ -1,9 +1,9 @@
 import { jsx } from '@emotion/core'
-import { Link } from '../components'
-import { Tag } from '../components/Tag'
-import { Text } from '../components/Text'
-import { palette, spacing } from '../../styles/theme'
-import { aboveTablet } from '../../styles/media'
+import { Link } from './Link'
+import { Tag } from './Tag'
+import { Text } from './Text'
+import { palette, spacing } from '../styles/theme'
+import { aboveTablet } from '../styles/media'
 import { Box } from './Box'
 /** @jsx jsx */ jsx
