+++ pages/blog.tsx
@@ -6,10 +6,11 @@ import { getPostBySlug } from '../lib/getPostBySlug'
 import { getPostFilePaths } from '../lib/getPostFilePaths'
 import { slugifyPost } from '../lib/slugifyPost'
 import { PostFrontmatter } from '../lib/types'
-import { Heading, Layout } from '../src/components'
-import { PostListItem } from '../src/components/PostListItem'
-import { Text } from '../src/components/Text'
-import { Box } from '../src/components/Box'
+import { Heading } from '../components/Heading'
+import { Layout } from '../components/Layout'
+import { PostListItem } from '../components/PostListItem'
+import { Text } from '../components/Text'
+import { Box } from '../components/Box'
 import { textMaxWidth } from '../styles/variables'
 /** @jsx jsx */ jsx
