+++ pages/blog/[slug].tsx
@@ -12,15 +12,16 @@ import { getPostBySlug } from '../../lib/getPostBySlug'
 import { getPostFilePaths } from '../../lib/getPostFilePaths'
 import { slugifyPost } from '../../lib/slugifyPost'
 import { PostFrontmatter } from '../../lib/types'
-import { Layout, Link } from '../../src/components'
-import { Box } from '../../src/components/Box'
-import { Callout } from '../../src/components/Callout'
-import { Img } from '../../src/components/Img'
-import { MarkdownWrapperStyles } from '../../src/components/Markdown'
-import { Text } from '../../src/components/Text'
+import { Layout } from '../../components/Layout'
+import { Link } from '../../components/Link'
+import { Box } from '../../components/Box'
+import { Callout } from '../../components/Callout'
+import { Img } from '../../components/Img'
+import { MarkdownWrapperStyles } from '../../components/Markdown'
+import { Text } from '../../components/Text'
 import { palette, spacing } from '../../styles/theme'
 import { container } from '../../styles/variables'
-import { Tag } from '../../src/components/Tag'
+import { Tag } from '../../components/Tag'
 /** @jsx jsx */ jsx
 
 interface PostPageParams extends ParsedUrlQuery {
