+++ pages/index.tsx
@@ -1,8 +1,10 @@
 import { css, jsx } from '@emotion/core'
 import React from 'react'
-import { Heading, Layout, Link } from '../src/components'
-import { Text } from '../src/components/Text'
-import { Box } from '../src/components/Box'
+import { Heading } from '../components/Heading'
+import { Layout } from '../components/Layout'
+import { Link } from '../components/Link'
+import { Text } from '../components/Text'
+import { Box } from '../components/Box'
 /** @jsx jsx */ jsx
 
 const navigation = [
