+++ pages/tags/[tag].tsx
@@ -4,13 +4,13 @@ import { getPostBySlug } from '../../lib/getPostBySlug'
 import { getPostFilePaths } from '../../lib/getPostFilePaths'
 import { slugifyPost } from '../../lib/slugifyPost'
 import { PostFrontmatter } from '../../lib/types'
-import { Box } from '../../src/components/Box'
-import { Heading } from '../../src/components/Heading'
-import { Layout } from '../../src/components/Layout'
-import { Text } from '../../src/components/Text'
+import { Box } from '../../components/Box'
+import { Heading } from '../../components/Heading'
+import { Layout } from '../../components/Layout'
+import { Text } from '../../components/Text'
 import { textMaxWidth } from '../../styles/variables'
 import { compareDesc, format } from 'date-fns'
-import { PostListItem } from '../../src/components/PostListItem'
+import { PostListItem } from '../../components/PostListItem'
 import { css, jsx } from '@emotion/core'
 /** @jsx jsx */ jsx
