--- src/components/index.js
@@ -1,5 +1,0 @@
-export { Heading } from './Heading'
-export { Link } from './Link'
-export { Markdown } from './Markdown'
-export { Section } from './Section'
-export { Layout } from './Layout'
